Missing needed returns
We need to return 0 for success. It causes troubles at parsing the yaml, cause it is checking for 0 for success, but we do not return 0, it could be any number/its undefined. Normally the compiler should warn you about that, but I guess to warning level wasnt high enough.


With this fix I'am now able to build https://github.com/yifanlu/taiHEN