Implement support for automatic symbol exports
vita-elf-create now supports the generation of an export configuration from symbol visibility information in the symbol table. This functionality is enabled by the option '-g <output-export-file>', which is used similarly to the option '-e <input-export-file>'.

Export generation can be used with the CMake toolchain through the use of the 'GEN_EXPORTS <output-export-file>' option for vita_create_self and the 'GEN_EXPORTS' option for vita_create_stubs.

Some compiler warnings were also fixed.